check_network: add performance data for links with available speed
When a link has a defined speed (returned by ioctl SIOCETHTOOL)
do add to the pluging output the min and max performance data in bytes.

Example:

  link enp1s0: 1Gbit/s (full duplex)
    --> enp1s0_txbyte/s=781;;;0;125000000
        enp1s0_rxbyte/s=1083;;;0;125000000

+++ include/netinfo.h
@@ -43,7 +43,7 @@ extern "C"
     int link_up;
     int link_running;
     unsigned int multicast;
-    unsigned long long speed;
+    unsigned int speed;		/* the link speed in Mbps */
     int duplex;
     struct iflist *next;
   } iflist_t;
